refactor(select): component optimization
