+++ packages/components/select/option-group.tsx
@@ -2,11 +2,11 @@ import { defineComponent, inject, computed } from 'vue';
 import { useTNodeJSX } from '../hooks/tnode';
 import props from './option-group-props';
 import { usePrefixClass, useCommonClassName } from '../hooks/useConfig';
-import { selectInjectKey } from './helper';
+import { selectInjectKey } from './consts';
 
 export default defineComponent({
   name: 'TOptionGroup',
-  props: { ...props },
+  props,
   setup(props) {
     const selectProvider = inject(selectInjectKey);
     const COMPONENT_NAME = usePrefixClass('select-option-group');
