+++ packages/components/select/select.tsx
@@ -17,9 +17,9 @@ import useDefaultValue from '../hooks/useDefaultValue';
 import useVModel from '../hooks/useVModel';
 import { useTNodeJSX } from '../hooks/tnode';
 import { useConfig, usePrefixClass } from '../hooks/useConfig';
-import { selectInjectKey, getSingleContent, getMultipleContent } from './helper';
-import { useSelectOptions } from './hooks/useSelectOptions';
-import useKeyboardControl from './hooks/useKeyboardControl';
+import { getSingleContent, getMultipleContent } from './utils';
+import { selectInjectKey } from './consts';
+import { useSelectOptions, useKeyboardControl } from './hooks';
 import type { PopupProps, PopupVisibleChangeContext } from '../popup';
 import type { SelectInputValueChangeContext } from '../select-input';
 import type { TdSelectProps, SelectValue } from './type';
